+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.2.1](https://kaos.sh/ek/13.2.1)
 
+- `[terminal/input]` Added `NewLine` flag
 - `[sliceutil]` Methods `Copy`, `IsEqual`, `Index`, `Contains`, and `Deduplicate` marked as deprecated
 - `[terminal/input]` Improved TMUX support